# docs/swagger.yaml
definitions:
  controllers.LoginInput:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  controllers.RegisterInput:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  controllers.bookInput:
    properties:
      description:
        type: string
      image_url:
        type: string
      price:
        type: string
      release_year:
        maximum: 2021
        minimum: 1980
        type: integer
      title:
        type: string
      total_page:
        type: integer
    required:
    - description
    - image_url
    - price
    - release_year
    - title
    - total_page
    type: object
  controllers.dosenInput:
    properties:
      mataKuliahID:
        type: integer
      nama:
        type: string
    required:
    - nama
    type: object
...
  models.Books:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      price:
        type: string
      release_year:
        type: integer
      thickness:
        type: string
      title:
        type: string
      total_page:
        type: integer
      updatedAt:
        type: string
    type: object
  models.Dosen:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      jadwalKuliah:
        items:
          $ref: '#/definitions/models.JadwalKuliah'
        type: array
      mataKuliahID:
        type: integer
      nama:
        type: string
      updatedAt:
        type: string
    type: object
  models.JadwalKuliah:
    properties:
      createdAt:
        type: string
      dosenID:
        type: integer
      hari:
        type: string
      id:
        type: integer
      jam_mulai:
        type: string
      jam_selesai:
        type: string
      mahasiswaID:
        type: integer
      updatedAt:
        type: string
    type: object
...
  models.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      role:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
  /auth/me:
    get:
      description: Get Current User by token.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      security:
      - BearerToken: []
      summary: Get Current User by token.
      tags:
      - Auth
  /books:
    get:
      description: Get a list of books.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Books'
            type: array
      summary: Get all books.
      tags:
      - book
    post:
      description: Creating a new book.
      parameters:
      - description: the body to create a new book
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.bookInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Books'
      security:
      - BearerToken: []
      summary: Create New book.
      tags:
      - book
  /books/{id}:
    delete:
      description: Delete a book by id.
      parameters:
      - description: book id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one book.
      tags:
      - book
    get:
      description: Get a book by id.
      parameters:
      - description: book id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Books'
      summary: Get book.
      tags:
      - book
    patch:
      description: Update book by id.
      parameters:
      - description: book id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: the body to update an book
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.bookInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Books'
      security:
      - BearerToken: []
      summary: Update book.
      tags:
      - book
  /dosen:
    get:
      description: Get a list of dosen.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Dosen'
            type: array
      summary: Get all Dosen.
      tags:
      - dosen
    post:
      description: Creating a new dosen.
      parameters:
      - description: the body to create a new dosen
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.dosenInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Dosen'
      security:
      - BearerToken: []
      summary: Create New dosen.
      tags:
      - dosen
  /dosen/{id}:
    delete:
      description: Delete a dosen by id.
      parameters:
      - description: dosen id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one dosen.
      tags:
      - dosen
    get:
      description: Get a dosen by id.
      parameters:
      - description: dosen id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Dosen'
      summary: Get dosen.
      tags:
      - dosen
    patch:
      description: Update dosen by id.
      parameters:
      - description: dosen id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update an dosen
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.dosenInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Dosen'
      security:
      - BearerToken: []
      summary: Update dosen.
      tags:
      - dosen
...
  /login:
    post:
      description: Logging in to get jwt token to access admin or user api by roles.
      parameters:
      - description: the body to login a user
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Login as as user.
      tags:
      - Auth
...
  /register:
    post:
      description: registering a user from public access.
      parameters:
      - description: the body to register a user
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.RegisterInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Register a user.
      tags:
      - Auth
swagger: "2.0"